Initial local exception refactoring
