+++ csharp/test/Ice/servantLocator/ServantLocatorI.cs
@@ -114,21 +114,15 @@ namespace Ice
                 }
                 else if (current.operation == "unknownUserException")
                 {
-                    var ex = new Ice.UnknownUserException();
-                    ex.unknown = "reason";
-                    throw ex;
+                    throw new Ice.UnknownUserException("reason");
                 }
                 else if (current.operation == "unknownLocalException")
                 {
-                    var ex = new Ice.UnknownLocalException();
-                    ex.unknown = "reason";
-                    throw ex;
+                    throw new Ice.UnknownLocalException("reason");
                 }
                 else if (current.operation == "unknownException")
                 {
-                    var ex = new Ice.UnknownException();
-                    ex.unknown = "reason";
-                    throw ex;
+                    throw new Ice.UnknownException("reason");
                 }
                 else if (current.operation == "userException")
                 {
